+++ core/src/com/template/game/state/update/UpdaterInterface.java
@@ -2,5 +2,5 @@ package com.template.game.state.update;
 
 public interface UpdaterInterface<T> {
 
-    void update(T object, float delta);
+    void update(T object, double delta);
 }
